requirements.txt
asgiref
crispy-bootstrap5
Django
django-crispy-forms
django-environ
django-filter
djangorestframework
Markdown
MouseInfo
pillow
psycopg2
PyAutoGUI
PyGetWindow
PyMsgBox
pyperclip
PyRect
PyScreeze
pytweening
sqlparse
tzdata
